services/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codelens-api"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "boto3>=1.35.0",
    "gitpython>=3.1.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "networkx>=3.4.0",
    "voyageai>=0.3.0",
    "qdrant-client>=1.12.0",
    "openai>=1.50.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = ["core*", "models*", "routers*", "services*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
